Add Akkadian LanguagePack.
Akkadian was a language broadly used in ancient Mesopotamia.
It takes its name from the early empire of Akkad, but was
also used in Babylon, Assyria, and other major states in
the region from approximately 2000 BCE to 100 CE.

The native script for the language is cuneiform. However,
this is a keyboard layout for the Akkadian language input
in latin normalization or transliteration.

As far as I can tell there isn't a standard layout for this,
so I've created one ad hoc, replacing unused letters with
the extra consonants derived from diacritics.

Vowel accents (macro and circumflex for normalization;
acute and grave for transliteration) are on the long-press
popups. Some variant letters are also on the popup, including
ayin and aleph on /a/.

No dictionary so far.