Akkadian: add superscript d.
Adding a full superscript alphabet is a bit hacky in unicode,
and requires code changes for another shift button. But the
most common determinative (or maybe it just feels like it)
is the DINGIR sign marking divine names.

Experiment with adding U+1D48 Modifier Letter Small D to the
secondary characters popup.